PROJECTM/requirements.txt
boto==2.6.0
certifi==2018.4.16
chardet==3.0.4
django==1.11.1
djangorestframework==3.8.2
elementtree===1.2.7-20070827-preview
html5lib==0.90
idna==2.7
markdown==2.0.3
mysql-python==1.2.3
psycopg2==2.6
pysqlite==2.6.3
python-openid==2.2.5
pytz==2017.2
requests==2.19.1
south==0.7.6
urllib3==1.23
virtualenv==1.11.2
virtualenvwrapper-win==1.1.5
